Fixed: Function WriteKeyVal() using prefix 0 on decimal values This can corrupt save files, since all properties will be saved as dec values but using hex format (prefix 0), so sphere won't read correctly these values on next reboot

+++ src/graysvr/CWorld.cpp
@@ -1779,7 +1779,7 @@ bool CWorld::LoadWorld() // Load world from script
 	CGString sDataName;
 	sDataName.Format("%s" GRAY_FILE "data",	static_cast<LPCTSTR>(g_Cfg.m_sWorldBaseDir));
 
-	unsigned int iPrevSaveCount = m_iSaveCountID;
+	long iPrevSaveCount = m_iSaveCountID;
 	for (;;)
 	{
 		LoadFile(sDataName, false);
