+++ src/graysvr/CWorld.h
@@ -378,10 +378,10 @@ public:
 	unsigned int m_SectorsQty;
 
 public:
-	unsigned int	m_iSaveCountID;	// Current archival backup id. Whole World must have this same stage id
-	unsigned int	m_iLoadVersion;		// Previous load version. (only used during load of course)
-	unsigned int	m_iPrevBuild;		// Previous __GITREVISION__
-	CServTime m_timeStartup;		// When did the system restore load/save ?
+	long m_iSaveCountID;		// Current archival backup id. Whole World must have this same stage id
+	long m_iLoadVersion;		// Previous load version. (only used during load of course)
+	long m_iPrevBuild;			// Previous __GITREVISION__
+	CServTime m_timeStartup;	// When did the system restore load/save ?
 
 	CGrayUID m_uidLastNewItem;	// for script access.
 	CGrayUID m_uidLastNewChar;	// for script access.
